+++ CHANGELOG.md
@@ -78,3 +78,4 @@
   - ENHANCEMENT: DiffEngine now gives a combined method to get removes, modifies then deletes in one call.
   - Ability to more easily set data in the back-end without using voodoo (the value type is derrived)
     (e.g session.set_data_by_xpath(context, xpath, value)
+  - FIX: Template conversion into XML from XPATH (list-values which looked like xpath were affecting the structure)